refactor(memory): add new pmm definition
