+++ include/memory/fpage.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "system/types.h"
+
 namespace task::ipc
 {
