+++ kern/memory/pmm/CMakeLists.txt
@@ -1,4 +1,8 @@
 cmake_minimum_required(VERSION 3.16 FATAL_ERROR)
 
 target_sources(kernel
-    PRIVATE buddy_pmm.cc kmem.cc pmm.cc)
+        PRIVATE buddy_pmm.cc
+        PRIVATE kmem.cc
+        PRIVATE pmm.cc
+        PRIVATE buddy_provider.cc
+        PRIVATE pmm_old.cc)
