+++ kern/memory/pmm/buddy_pmm.cc
@@ -1,4 +1,4 @@
-#include "./buddy_pmm.h"
+#include "include/buddy_pmm.h"
 
 #include "system/memlayout.h"
 #include "system/pmm.h"
